+++ torch_spyre/_inductor/passes.py
@@ -65,7 +65,7 @@ from .work_division import (
 from .pass_utils import apply_splits_from_index_coeff, iteration_space_from_op
 from .scratchpad.allocator import (
     StrategyBCoOptimizingAllocator,
-    IlpCoOptimizingAllocator,
+    CoOptimizingAllocator,
     scratchpad_planning,
 )
 from .fusion import spyre_fuse_nodes
@@ -291,16 +291,14 @@ def _maybe_scratchpad_planning(graph: GraphLowering) -> None:
     if not config.lx_planning:
         return
     # Pick the allocator by solver: the ILP solver runs its own joint
-    # core-division solve, so layout_solver == "ilp" uses IlpCoOptimizingAllocator
+    # core-division solve, so layout_solver == "ilp" uses CoOptimizingAllocator
     # regardless of the co-opt flag. StrategyB is the co-optimizing allocator for
     # the gap-based (greedy/bestfit/firstfit) solvers and is used only when the
     # co-opt flag is set and the solver is not ILP; otherwise allocator stays None
     # (the DefaultAllocator placement-only path).
-    allocator: Optional[IlpCoOptimizingAllocator | StrategyBCoOptimizingAllocator] = (
-        None
-    )
+    allocator: Optional[CoOptimizingAllocator | StrategyBCoOptimizingAllocator] = None
     if config.layout_solver == "ilp":
-        allocator = IlpCoOptimizingAllocator()
+        allocator = CoOptimizingAllocator()
     elif config.co_optimizing_lx_planning:
         allocator = StrategyBCoOptimizingAllocator()
     scratchpad_planning(graph, allocator=allocator)
